requirements.txt
requests
pandas 
yagmail 
python-decouple